+++ test/dummy/app/controllers/pages_controller.rb
@@ -1,6 +1,4 @@
 class PagesController < ApplicationController
-  before_action :skip_authorization, :skip_policy_scope
-
   def show
     render "/pages/#{path}"
   end